+++ src/test/java/com/caoccao/javet/swc4j/ast/expr/TestSwc4jAstBinExpr.java
@@ -26,6 +26,7 @@ import com.caoccao.javet.swc4j.ast.visitors.Swc4jAstVisitorResponse;
 import com.caoccao.javet.swc4j.exceptions.Swc4jCoreException;
 import com.caoccao.javet.swc4j.outputs.Swc4jParseOutput;
 import com.caoccao.javet.swc4j.plugins.ISwc4jPluginHost;
+import com.caoccao.javet.swc4j.utils.SimpleList;
 import com.caoccao.javet.swc4j.utils.SimpleMap;
 import org.junit.jupiter.api.Test;
 
@@ -96,6 +97,41 @@ public class TestSwc4jAstBinExpr extends BaseTestSuiteSwc4jAst {
             assertEquals("a", binExpr.getLeft().as(Swc4jAstIdent.class).getSym());
             assertEquals("b", binExpr.getRight().as(Swc4jAstIdent.class).getSym());
             assertEquals(bangCount, binExpr.getBangCount());
+        }
+    }
+
+    @Test
+    public void testGetLogicalOperatorCount() throws Swc4jCoreException {
+        Map<String, List<Object>> testCaseMap = SimpleMap.of(
+                "a==b", SimpleList.of(
+                        0, Swc4jAstBinaryOp.EqEq),
+                "a>b||c<d", SimpleList.of(
+                        0, Swc4jAstBinaryOp.LogicalOr,
+                        1, Swc4jAstBinaryOp.Gt,
+                        1, Swc4jAstBinaryOp.Lt),
+                "a>b||(c==d&&(e+f)<=g)", SimpleList.of(
+                        0, Swc4jAstBinaryOp.LogicalOr,
+                        1, Swc4jAstBinaryOp.Gt,
+                        1, Swc4jAstBinaryOp.LogicalAnd,
+                        2, Swc4jAstBinaryOp.EqEq,
+                        2, Swc4jAstBinaryOp.LtEq,
+                        3, Swc4jAstBinaryOp.Add),
+                "a>b||((c+d)?e:f)", SimpleList.of(
+                        0, Swc4jAstBinaryOp.LogicalOr,
+                        1, Swc4jAstBinaryOp.Gt,
+                        0, Swc4jAstBinaryOp.Add));
+        for (Map.Entry<String, List<Object>> entry : testCaseMap.entrySet()) {
+            String code = entry.getKey();
+            int size = entry.getValue().size() >> 1;
+            Swc4jParseOutput output = swc4j.parse(code, tsScriptParseOptions);
+            List<Swc4jAstBinExpr> nodes = output.getProgram().find(Swc4jAstBinExpr.class);
+            assertEquals(size, nodes.size());
+            for (int i = 0; i < size; i++) {
+                Swc4jAstBinExpr binExpr = nodes.get(i);
+                int index = i << 1;
+                assertEquals(entry.getValue().get(index + 1), binExpr.getOp());
+                assertEquals(entry.getValue().get(index), binExpr.getLogicalOperatorCount());
+            }
         }
     }