+++ docs/release_notes.md
@@ -4,7 +4,7 @@
 
 * Added `find()` to `ISwc4jAst`
 * Added `getMinusCount()` to `Swc4jAstNumber`
-* Added `getBangCount()` to `Swc4jAstBinExpr`
+* Added `getBangCount()`, `getLogicalOperatorCount()` to `Swc4jAstBinExpr`
 * Added `getOppositeOperator()`, `isLogicalOperator()`, `isLogicalCompareOperator()`, `isLogicalConditionOperator()` to `Swc4jAstBinaryOp`
 
 ## 1.2.0